combat_package/pyproject.toml
[project]
name = "worldseed-combat"
version = "0.1.0"
description = "WorldSeed combat engine (standalone package)."
requires-python = ">=3.12"
dependencies = ["PyYAML>=6.0"]
[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "black", "ruff", "mypy", "pre-commit"]
[tool.pytest.ini_options]
addopts = "-q"
[tool.black]
target-version = ["py312"]
line-length = 100
[tool.ruff]
line-length = 100
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
